// src/app/shared/enums.ts
export enum EventType {
  Adopt = "Adopt",
  Date = "Date",
  Birth = "Birth",
  BreakUp = "Break Up",
  Divorce = "Divorce",
  Marriage = "Marriage"
}

export enum Stage { 
  Baby,
  Toddler,
  Child,
  Teenager,
  YoungAdult,
  Adult,
  Elder
}

export enum Place {
  NotInTheWorld = "Not in the World",
  BrindletonBay = "Brindleton Bay",
  Britechester = "Britechester",
  DelSolValley = "Del Sol Valley",
  ForgottenHollow = "Forgotten Hollow",
  Glimmerbrook = "Glimmerbrook",
  GraniteFalls = "Granite Falls",
  Newcrest = "Newcrest",
  OasisSprings = "Oasis Springs",
  SanMyshuno = "San Myshuno",
  Strangeville = "Strangerville",
  Sulani = "Sulani",
  WillowCreek = "Willow Creek",
  Windenburg = "Windenburg"
}

export enum Career {
  Astronaut = "Astronaut",
  Athletic = "Athletic",
  Babysitter = "Babysitter",
  Barista = "Barista",
  Business = "Business",
  Conservationist = "Conservationist",
  Criminal = "Criminal",
  Critic = "Critic",
  Culinary = "Culinary",
  Diver = "Diver",
  DramaClub = "Drama Club",
  Education = "Education",
  Engineer = "Engineer",
  Entertainment = "Entertainment",
  ESports = "E-Sports Competitor",
  FastFood = "Fast Food Employee",
  Film = "Film",
  Fisherman = "Fisherman",
  Freelancer = "Freelancer",
  Gardener = "Gardener",
  Law = "Law",
  LawEnforcement = "Law Enforcement",
  Lifeguard = "Lifeguard",
  ManualLaborer = "Manual Laborer",
  Medicine = "Medicine",
  Military = "Military",
  Painter = "Painter",
  Politics = "Politics",
  RetailEmployee = "Retail Employee",
  Science = "Science",
  Scout = "Scout",
  SecretAgent = "Secret Agent",
  SoccerTeamPlayer = "Soccer Team Player",
  SocialMedia = "Social Media",
  StyleInfluencer = "Style Influencer",
  TechGuru = "Tech Guru",
  Writer = "Writer",
  Unemployed = "Unemployed"
}